fix: wrap command reg in useEffect
