+++ src/components/plugins/TaskPlugin/TaskNode.tsx
@@ -1,17 +1,11 @@
 import {
-  createCommand,
   type EditorConfig,
   type LexicalNode,
   type NodeKey,
   SerializedLexicalNode,
   LexicalEditor,
   DecoratorNode,
-  $insertNodes,
-  COMMAND_PRIORITY_NORMAL,
 } from "lexical"
-import { $insertNodeToNearestRoot } from "@lexical/utils"
-import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext"
-import { $wrapNodes } from "@lexical/selection"
 
 import { TaskComponent } from "./TaskComponent"
 
@@ -20,6 +14,9 @@ export type SerializedTaskNode = {
   desciption?: string
 } & SerializedLexicalNode
 
+/**
+ * Custom task node
+ */
 export class TaskNode extends DecoratorNode<JSX.Element> {
   __title?: string
   __description?: string
@@ -74,17 +71,6 @@ export class TaskNode extends DecoratorNode<JSX.Element> {
     return false
   }
 
-  // insertNewAfter(
-  //   _selection: RangeSelection,
-  //   restoreSelection?: boolean | undefined
-  // ): LexicalNode | null {
-  //   const newBlock = $createParagraphNode()
-  //   const direction = this.getDirection()
-  //   newBlock.setDirection(direction)
-  //   this.insertAfter(newBlock, restoreSelection)
-  //   return newBlock
-  // }
-
   exportJSON(): SerializedTaskNode {
     return {
       title: this.getTitle(),
@@ -104,25 +90,6 @@ export function $createTaskNode(): TaskNode {
   return new TaskNode()
 }
 
-export function $isBannerNode(node: LexicalNode): node is TaskNode {
+export function $isTaskNode(node: LexicalNode): node is TaskNode {
   return node instanceof TaskNode
-}
-
-export const INSERT_BANNER_COMMAND = createCommand("insertTask")
-
-export function TaskPlugin(): null {
-  const [editor] = useLexicalComposerContext()
-  if (!editor.hasNodes([TaskNode])) {
-    throw new Error("TaskPlugin: TaskPlugin not registered on editor")
-  }
-  editor.registerCommand(
-    INSERT_BANNER_COMMAND,
-    () => {
-      const taskNode = $createTaskNode()
-      $insertNodeToNearestRoot(taskNode)
-      return true
-    },
-    COMMAND_PRIORITY_NORMAL
-  )
-  return null
 }
